requirements.txt
Django==4.2.10
django-crispy-forms==2.0
Pillow==10.4.0
qrcode==7.4.2
xlwt==1.3.0
openpyxl==3.1.3
pypng==0.20220715.0
django-mssql-backend==2.8.1
pyodbc==4.0.39 